+++ dune/common/test/fvectortest.cc
@@ -10,6 +10,7 @@
 #include <typeinfo>
 #include <type_traits>
 
+#include <dune/common/bigfloat.hh>
 #include <dune/common/classname.hh>
 #include <dune/common/exceptions.hh>
 #include <dune/common/fvector.hh>
@@ -589,6 +590,19 @@ int main()
     FieldVectorTest<float, 3>();
     FieldVectorTest<double, 3>();
     FieldVectorTest<long double, 3>();
+#if HAVE_MPFR
+    {
+      typedef Dune::BigFloat<128u> ft;
+      FieldVectorMainTest<ft,ft,3>();
+      FieldVectorMainTest<ft,ft,2>();
+      FieldVectorMainTest<ft,ft,1>();
+      FieldVectorMainTest<ft,ft,0>();
+      ScalarOperatorTest<ft>();
+      ScalarOrderingTest<ft>();
+      DotProductTest<ft,3>();
+    }
+#endif // HAVE_GMP
+
 #if HAVE_GMP
     {
       // we skip the complex test and the int test, as these will be very hard to implement with GMPField